% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ccfrp_otolith.R
\name{clean_ccfrp_otolith}
\alias{clean_ccfrp_otolith}
\title{Format otolith data}
\usage{
clean_ccfrp_otolith(data, species, year)
}
\arguments{
\item{data}{add definition}

\item{species}{add definition}

\item{year}{add definition}
}
\description{
Format otolith data
}
\author{
Chantel Wetzel
}
