% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_cdfw_otolith_files.R
\name{clean_cdfw_otolith_files}
\alias{clean_cdfw_otolith_files}
\title{Format otolith data}
\usage{
clean_cdfw_otolith_files(data, species, year)
}
\arguments{
\item{data}{add definition}

\item{species}{add definition}

\item{year}{add definition}
}
\description{
Format otolith data
}
\author{
Chantel Wetzel
}
