% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_gcdc_data.R
\name{clean_gcdc_data}
\alias{clean_gcdc_data}
\title{Clean and standardize CCFRP data}
\usage{
clean_gcdc_data(data, species, year = 2000)
}
\arguments{
\item{data}{data frame}

\item{species}{add definition}

\item{year}{year to filter the data}
}
\description{
Clean and standardize CCFRP data
}
\author{
Chantel Wetzel
}
