% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_recfin_ages.R
\name{clean_recfin_ages}
\alias{clean_recfin_ages}
\title{Clean RecFIN data}
\usage{
clean_recfin_ages(species, data, year)
}
\arguments{
\item{species}{add definition}

\item{data}{add definition}

\item{year}{add definition}
}
\description{
Clean RecFIN data
}
\author{
Chantel Wetzel
}
