% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_recfin_lengths.R
\name{clean_recfin_lengths}
\alias{clean_recfin_lengths}
\title{Clean RecFIN length data}
\usage{
clean_recfin_lengths(data, species, year)
}
\arguments{
\item{species}{add definition}

\item{year}{add definition}

\item{or_data}{add definition}

\item{wa_data}{add definition}

\item{ca_data}{add definition}
}
\description{
Clean RecFIN length data
}
\author{
Chantel Wetzel
}
