% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_wdfw_otolith_files.R
\name{clean_wdfw_otolith_files}
\alias{clean_wdfw_otolith_files}
\title{Format otolith data}
\usage{
clean_wdfw_otolith_files(data, species, year)
}
\arguments{
\item{data}{add definition}

\item{source}{add definition}

\item{state}{add definition}
}
\description{
Format otolith data
}
\author{
Chantel Wetzel
}
