% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_all_data.R
\name{combine_all_data}
\alias{combine_all_data}
\title{Combine all data into a data frame}
\usage{
combine_all_data(dir = dir, wcgbt, nwfsc_hkl)
}
\arguments{
\item{dir}{Directory location to save the combined data frame}

\item{wcgbt}{add definition}

\item{nwfsc_hkl}{add definition}
}
\description{
Combine all data into a data frame
}
\author{
Chantel Wetzel
}
