% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_short_species_list.R
\name{get_short_species_list}
\alias{get_short_species_list}
\title{List of select species to summarize commercial and recreational data for}
\usage{
get_short_species_list()
}
\description{
List of select species to summarize commercial and recreational data for
}
\author{
Chantel Wetzel
}
