% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rdata_object.R
\name{load_rdata_object}
\alias{load_rdata_object}
\title{Load rdata file}
\usage{
load_rdata_object(dir, rdata_name)
}
\arguments{
\item{dir}{Directory where a rdata object is located}

\item{rdata_name}{file name}
}
\description{
Load rdata file
}
\author{
Chantel Wetzel
}
