% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data_by_year.R
\name{plot_data_by_year}
\alias{plot_data_by_year}
\title{Plot all the data}
\usage{
plot_data_by_year(dir = here::here("plots", "state_comparisons"), data)
}
\arguments{
\item{dir}{Directory location to save the combined data frame}

\item{data}{add definition}

\item{year}{add definition}
}
\description{
Plot all the data
}
\author{
Chantel Wetzel
}
