% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hkl_comps.R
\name{plot_hkl_comps}
\alias{plot_hkl_comps}
\title{Plot NWFSC HKL data}
\usage{
plot_hkl_comps(dir = here::here("plots", "hkl_comps"), data)
}
\arguments{
\item{dir}{add definition}

\item{data}{add definition}
}
\description{
Plot NWFSC HKL data
}
\author{
Chantel Wetzel
}
