% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_wcgbt_species.R
\name{rename_wcgbt_species}
\alias{rename_wcgbt_species}
\title{Clean NWFSC WCGBTS catch and biological data}
\usage{
rename_wcgbt_species(data)
}
\arguments{
\item{data}{Dataframe of WCGBT data}
}
\description{
Clean NWFSC WCGBTS catch and biological data
}
\author{
Chantel Wetzel
}
